Format and auto load
